+++ src/shared/ui/calendar/index.ts
@@ -4,7 +4,7 @@ export { default as HeaderButton } from './ui/header-button'
 export { default as HeaderContainer } from './ui/header-container'
 export { default as HeaderMonthLabel } from './ui/header-month-label'
 export { default as InfiniteCalendar } from './ui/infinite-calendar'
-export { default as MonthlyCalendar } from './ui/montly-calendar'
+export { default as MonthlyCalendar } from './ui/monthly-calendar'
 export { default as PrivateScheduleBadge } from './ui/private-schedule-badge'
 export { default as PrivateScheduleModal } from './ui/private-schedule-modal'
 export { default as AddScheduleButton } from './ui/add-schedule-button'
